% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{beating}
\alias{beating}
\title{This beating definition that I made up is just the total energy per cycle}
\usage{
beating(x)
}
\description{
This beating definition that I made up is just the total energy per cycle
}
