% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{compute_cycle_length}
\alias{compute_cycle_length}
\title{Compute the cycle length of a complex wave}
\usage{
compute_cycle_length(x, ref, dimension)
}
\arguments{
\item{x}{Spectrum representing a complex waveform}

\item{dimension}{Space or time, used to label the output}
}
\value{
Estimated cycle length of the complex waveform.
}
\description{
Compute the cycle length of a complex wave
}
