% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_energy_per_cycle}
\alias{compute_energy_per_cycle}
\title{Compute beating perception from beat spectrum}
\usage{
compute_energy_per_cycle(x)
}
\arguments{
\item{x}{The beat spectrum, containing frequency and amplitude information for beats derived from the stimulus.}
}
\value{
Psychophysical measure of beating perception.
}
\description{
Calculates a psychophysical measure of beating perception derived from the
beat spectrum, interpreting beat frequencies and amplitudes in a perceptual context.
}
