% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_fundamental_beats_wavenumber}
\alias{compute_fundamental_beats_wavenumber}
\title{Compute the fundamental beat of the complex waveform.}
\usage{
compute_fundamental_beats_wavenumber(x)
}
\arguments{
\item{x}{Wavelength spectrum that include stimulus, beat, and cochlear emission tones.}
}
\value{
Fundamental wavenumber of a complex waveform.
}
\description{
Computes the fundamental beat from a beat spectrum.
}
