% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_fundamental_wavenumber}
\alias{compute_fundamental_wavenumber}
\title{Compute the fundamental wavenumber of the complex waveform.}
\usage{
compute_fundamental_wavenumber(x)
}
\arguments{
\item{x}{Wavelength spectrum that include stimulus, beat, and cochlear emission tones.}
}
\value{
Fundamental wavenumber of a complex waveform.
}
\description{
Computes the fundamental wavenumber from a wavelength spectrum that
includes stimulus, beat, and cochlear emission tones.
}
