% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{compute_space_octave_cycles}
\alias{compute_space_octave_cycles}
\title{Compute the spatial octave cycle length of the complex waveform.}
\usage{
compute_space_octave_cycles(x)
}
\arguments{
\item{x}{Wavelength spectrum that include stimulus, side bands.}
}
\value{
spatial octave cycle length
}
\description{
Computes the spatial octave cycle length from a wavelength spectrum that
includes stimulus, side bands.
}
