% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{compute_space_pseudo_octave}
\alias{compute_space_pseudo_octave}
\title{Compute the spatial pseudo octave of the complex waveform.}
\usage{
compute_space_pseudo_octave(x)
}
\arguments{
\item{x}{Wavelength spectrum that include stimulus, side bands.}
}
\value{
spatial pseudo octave
}
\description{
Computes the spatial pseudo octave from a wavelength spectrum that
includes stimulus, side bands.
}
