% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{compute_time_cycles}
\alias{compute_time_cycles}
\title{Compute the temporal cycle length of a complex waveform.}
\usage{
compute_time_cycles(x)
}
\arguments{
\item{x}{Wavelength spectrum that include stimulus and side bands.}
}
\value{
temporal cycle length
}
\description{
Computes the temporal cycle length from a frequency spectrum that
includes stimulus and side bands.
}
