% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{compute_time_pseudo_octave}
\alias{compute_time_pseudo_octave}
\title{Compute the temporal pseudo octave of the complex waveform.}
\usage{
compute_time_pseudo_octave(x)
}
\arguments{
\item{x}{Frequency spectrum that include stimulus, side bands.}
}
\value{
temporal pseudo octave
}
\description{
Computes the temporal pseudo octave from a frequency spectrum that
includes stimulus, side bands.
}
