% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{energy_per_cycle}
\alias{energy_per_cycle}
\title{This beating definition that I made up is just the total energy per cycle}
\usage{
energy_per_cycle(x)
}
\description{
This beating definition that I made up is just the total energy per cycle
}
