% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{generate_distortion_products}
\alias{generate_distortion_products}
\title{Generate distortion products}
\usage{
generate_distortion_products(x)
}
\arguments{
\item{x}{A list or tibble containing \code{stimulus_frequency_spectrum}
(a data.frame/tibble with \code{frequency} and \code{amplitude} columns).}
}
\value{
The input \code{x} augmented with:
}
\description{
Generate distortion products
}
