% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{generate_beats}
\alias{generate_beats}
\title{Generate beat tones as a wavelength spectrum (only, not frequency spectrum)
and apply a pass filter (low, high, all, or none) relative to the lowest tone
in the stimulus.}
\usage{
generate_beats(x)
}
\arguments{
\item{x}{Wavelength and frequency spectra representing the stimulus tones.}
}
\value{
The wavelength spectrum of the selected beat tones.
}
\description{
Generate beat tones as a wavelength spectrum (only, not frequency spectrum)
and apply a pass filter (low, high, all, or none) relative to the lowest tone
in the stimulus.
}
