% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{generate_cochlea_amplifications}
\alias{generate_cochlea_amplifications}
\title{Generate tones emitted from the cochlea in response to frequency stimulation}
\usage{
generate_cochlea_amplifications(x)
}
\arguments{
\item{x}{Wavelength and frequency spectra representing the stimulus and beat tones.}
}
\value{
Combined frequency and wavelength spectra for the stimulus, beat tones, and tones generated by the cochlear amplification.
}
\description{
Generate tones emitted from the cochlea in response to frequency stimulation
}
