% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{generate_combination_tones}
\alias{generate_combination_tones}
\title{Generate beat tones}
\usage{
generate_combination_tones(x)
}
\arguments{
\item{x}{Wavelength and frequency spectra representing the stimulus tones.}
}
\value{
The wavelength spectrum of the selected beat tones.
}
\description{
Generate beat tones
}
