% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{generate_cubic_distortion_products}
\alias{generate_cubic_distortion_products}
\title{Generate generate_cubic_distortion_products}
\usage{
generate_cubic_distortion_products(x)
}
\arguments{
\item{x}{A list or tibble containing \code{stimulus_frequency_spectrum}
(a data.frame/tibble with \code{frequency} and \code{amplitude} columns).}
}
\value{
The input \code{x} augmented with:
\itemize{
\item \code{cubic_distortion_frequency_spectrum}: filtered sideband frequencies & amplitudes
\item \code{cubic_distortion_wavenumber_spectrum}: corresponding wavenumbers & amplitudes
}
}
\description{
For a given stimulus spectrum, compute the difference frequncies
and the spectral sidebands (fi ± |fj – fi|) for every unordered
pair of input frequencies, then convert to wavenumber and filter into range.
}
