% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{generate_sidebands}
\alias{generate_sidebands}
\title{Generate beat tones and sidebands}
\usage{
generate_sidebands(x)
}
\arguments{
\item{x}{Wavelength and frequency spectra representing the stimulus tones.}
}
\value{
The wavelength spectrum of the selected beat tones.
}
\description{
Generate beat tones and sidebands
}
