% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R-package.R
\docType{package}
\name{mami.codi.beaty.R-package}
\alias{mami.codi.beaty.R}
\alias{mami.codi.beaty.R-package}
\title{mami.codi.beaty.R: An Auditory Model of Major-Minor, Consonance-Dissonance and Beating Perception}
\description{
Allows researchers to explore music perception and music theory.
}
\author{
\strong{Maintainer}: Brian Mulloy \email{brian@homeymusic.com} (\href{https://orcid.org/0000-0002-1803-3172}{ORCID})

}
\keyword{internal}
