% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{speed_of_sound}
\alias{speed_of_sound}
\title{Speed of Sound}
\usage{
speed_of_sound()
}
\description{
Approximate speed of sound at sea level m / s
}
\details{
using E4 numerical value so we get some nice wavelength ratios but close to
value at room temp at sea level.
'
}
