% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stern_brocot}
\alias{stern_brocot}
\title{stern_brocot}
\usage{
stern_brocot(x, uncertainty, metadata = "")
}
\arguments{
\item{x}{Number to convert to rational fraction}

\item{uncertainty}{Binary search stops once the desired uncertainty is reached}
}
\value{
A ratio of num / den
}
\description{
Approximates a floating-point number to arbitrary uncertainty.
}
