% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{uncertainty}
\alias{uncertainty}
\title{Compute uncertainty}
\usage{
uncertainty()
}
\value{
Uncertainty
}
\description{
Uncertainty is the Heisenberg uncertainty limit divided by number of periods
}
