% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_parsing.R
\name{validate_combine_spectra}
\alias{validate_combine_spectra}
\title{Helper functions to combine spectra}
\usage{
validate_combine_spectra(..., tolerance = 1e-15)
}
\description{
Helper functions to combine spectra
}
