% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_energy_per_cycle}
\alias{compute_energy_per_cycle}
\title{Compute energy density per cycle}
\usage{
compute_energy_per_cycle(x)
}
\arguments{
\item{x}{The spectrum of stimulus and side bands.}
}
\value{
Physics measure of energy per cycle.
}
\description{
Calculates energy density per cycle.
}
