% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{compute_space_cycles}
\alias{compute_space_cycles}
\title{Compute the spatial cycle length of the complex waveform.}
\usage{
compute_space_cycles(x)
}
\arguments{
\item{x}{Wavelength spectrum that include stimulus, side bands.}
}
\value{
spatial cycle length
}
\description{
Computes the spatial cycle length from a wavelength spectrum that
includes stimulus, side bands.
}
