% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{energy_per_cycle}
\alias{energy_per_cycle}
\title{this is a made up metric that felt right to me
over one spatial cycle we are spatially
displacing material by amplitude A.}
\usage{
energy_per_cycle(x)
}
\description{
if we were to multiply it by bulk modulus
we would have units of energy
}
\details{
https://en.wikipedia.org/wiki/Bulk_modulus
}
