% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.R
\name{generate_beats}
\alias{generate_beats}
\title{Generate beat tones.}
\usage{
generate_beats(x)
}
\arguments{
\item{x}{Wavelength and frequency spectra representing the stimulus tones.}
}
\value{
The wavelength and frequency spectrum of the selected beat tones.
}
\description{
Generate beat tones.
}
