% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mami.codi.beaty.R
\name{generate_superposition_envelope}
\alias{generate_superposition_envelope}
\title{Generate amplitude modulation}
\usage{
generate_superposition_envelope(x)
}
\arguments{
\item{x}{A list or tibble containing \code{stimulus_frequency_spectrum}
(a data.frame/tibble with \code{frequency} and \code{amplitude} columns).}
}
\value{
The input \code{x} augmented with:
\itemize{
\item \code{amplitude_modulation_frequency_spectrum}: filtered sideband frequencies & amplitudes
\item \code{amplitude_modulation_wavelength_spectrum}: corresponding wavelengths & amplitudes
}
}
\description{
For a given stimulus spectrum, compute the difference frequncies
and the spectral sidebands (fi ± |fj – fi|) for every unordered
pair of input frequencies, then convert to wavelength and filter into range.
}
